requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3
google-generativeai>=0.4.0
ortools>=9.8
numpy>=1.24.0
python-dotenv>=1.0.0
